+++ src/pages/detailed-schedule/my-detailed-schedule/my-detailed-schedule-list-page.tsx
@@ -1,6 +1,6 @@
 import { Suspense } from 'react'
 
-import { MyDetailedScheduleList } from '@/features/my-schedule'
+import { MyDetailedScheduleList } from '@/features/my-schedule/ui'
 
 import DetailedScheduleListSkeleton from '../detailed-schedule-list-skeleton'
