+++ src/pages/home/home.tsx
@@ -4,7 +4,7 @@ import { Outlet } from 'react-router'
 import { useUserStore } from '@/entities/user/models/use-user-store'
 import EmailSigninModal from '@/features/auth/signin/ui/email-signin-modal'
 import SignupModal from '@/features/auth/signup/ui/signup-modal'
-import { MyCalendar } from '@/features/my-schedule'
+import { MyCalendar } from '@/features/my-schedule/ui'
 import { IconCalendarArrowLeft } from '@/shared/assets/icons'
 import useIntroGuide from '@/shared/hooks/use-intro-guide'
 import { Calendar, HeaderButton, HeaderContainer, HeaderMonthLabel, InfiniteCalendar } from '@/shared/ui'
